requirements.txt
Flask
python-dotenv
python-mysql-connector

